Handled exceptions and added hot potato books to the sum
Handled some exceptions where if no hot potato books, enchants or gems used on the item then the item will have no info about it at all, added a way to check if it's there in the first place before adding them to the calculation. Also added potato book to the sum and print format.